+++ conanfile.py
@@ -26,7 +26,7 @@ class RaftCoreGRPCConan(ConanFile):
             "lzma/5.2.4@bincrafters/stable",
             "raft_core/2019.01.24@oss/testing",
             "sds_grpc/1.1.0@sds/testing",
-            "sds_logging/3.8.1@sds/testing"
+            "sds_logging/3.8.2@sds/testing"
         )
 
     generators = "cmake"