+++ test_package/run_tests.sh
@@ -3,7 +3,7 @@
 set -eu
 export ASAN_OPTIONS=detect_leaks=0
 LOG_MODS="raft_core:2"
-CLIENT_VERBOSITY=1
+CLIENT_VERBOSITY=3
 SERVER_VERBOSITY=1
 
 CLIENT_CLI_OPTS="-csv ${CLIENT_VERBOSITY} --log_mods ${LOG_MODS}"
@@ -28,22 +28,31 @@ done
 
 echo "[          ] Settling members"
 sleep 5
+echo "[       OK ]"
 
 for i in $(seq 1 $((${SERVER_COUNT} - 2)))
 do
   echo "[          ] Adding Server $i"
   ./bin/raft_client ${CLIENT_CLI_OPTS} --add $i --server 0
   sleep 1
+  echo "[       OK ]"
 done
 
 echo "[          ] Settling group"
 
 echo "[          ] Writing Message"
 ./bin/raft_client ${CLIENT_CLI_OPTS} -m 'test::message' --server 1
+echo "[       OK ]"
 echo "[          ] Adding Server $((${SERVER_COUNT} - 1))"
 ./bin/raft_client ${CLIENT_CLI_OPTS} --add $((${SERVER_COUNT} - 1)) --server 2
 echo "[          ] Letting new member sync"
 sleep 5
+echo "[       OK ]"
+echo "[          ] Removing member from group"
+./bin/raft_client ${CLIENT_CLI_OPTS} --remove $((${SERVER_COUNT} - 2)) --server 2
+sleep 1
+echo "[       OK ]"
+
 echo "[          ] Counting Server Processes"
 alive_servers="$(ps -ef | grep -E '\./bin/raft_server' | grep -v grep | wc -l)"