+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### None
+### Added
+- Added test to remove server from group.
+
+### Fixed
+- Fixed remove server factory function.
 
 ## [0.10.0] - 2019.01.28
 ### Changed